requirements.txt
click==8.0.2
colorama==0.4.4
Flask==2.0.2
Flask-WTF==0.15.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
Werkzeug==2.0.2
WTForms==2.3.3
